breaking out executable to it's on file, also apparently there isn't as much es6 in node4 as I thought there was :-/
